+++ docs/content/public/llms.txt
@@ -26,7 +26,7 @@ Two cross-cutting conventions to know before reading anything else:
 - [Badge](https://ui.frappe.io/docs/components/badge): status pill; same `variant` + `theme` axes as Button.
 - [Breadcrumbs](https://ui.frappe.io/docs/components/breadcrumbs): page-header trail; takes `items: { label, route }[]`.
 - [Button](https://ui.frappe.io/docs/components/button): primary trigger element; supports `route`/`link` for nav, `icon`/`iconLeft`/`iconRight`, `loading`, `disabled`.
-- [Calendar](https://ui.frappe.io/docs/components/calendar): month/week calendar view.
+- [Calendar](https://ui.frappe.io/docs/experimental/calendar): month/week calendar view (experimental).
 - [Charts](https://ui.frappe.io/docs/components/charts): chart family (line, bar, etc.).
 - [Checkbox](https://ui.frappe.io/docs/components/checkbox): boolean input; standard labeling contract.
 - [Combobox](https://ui.frappe.io/docs/components/combobox): single-select with search; `v-model` value + `v-model:query`.
