+++ experimental.ts
@@ -18,6 +18,10 @@ export type {
   CodeEditorEmits,
   CodePreviewProps,
 } from './experimental/CodeEditor'
+// Calendar family. Moved out of root (#1020, redirect of #989) — parked
+// here, unstable, with its public API unchanged, until a redesigned
+// calendar family replaces it. Migration is the import-path change only.
+export * from './experimental/Calendar'
 // ListView family. Moved out of root (#985) — `frappe-ui/list` is
 // composition-based by design (P3) and doesn't replicate ListView's
 // config-driven columns (resizable widths, per-column getLabel/prefix
